testing up and running, still need to write tests
