+++ Pipfile
@@ -10,6 +10,7 @@ xmltodict = "*"
 ostruct = "*"
 gunicorn = "*"
 requests = "*"
+pytest = "*"
 
 [dev-packages]
