+++ run.py
@@ -1,0 +1,6 @@
+from stackoverflow import jobs
+
+app = jobs.create_app()
+
+if __name__ == "__main__":
+  app.run()
